Consolidate the two end-to-end verify gates into one, and harden JSON parsing
protean verify --json now returns typed Any from raw_decode; narrow it with
an isinstance check so a non-object envelope fails with a clear message
instead of a confusing KeyError downstream.

The new scaffold-verify e2e test duplicated the one already in
TestScaffoldVerifiesGreen (added by #1388): both generate a scaffold, run
protean verify --json as a subprocess, and assert the same stage statuses.
Running the expensive scaffold generation and verify pass twice slows the
suite for no extra coverage and risks drift between the two copies. Folded
the newer test's hardening (errors="replace", stray-stdout-tolerant parsing,
per-stage assertions) into the existing test and removed the duplicate.
